requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.4.5.1
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
dask==2021.5.0
distlib==0.3.1
fastapi~=0.68.2
filelock==3.0.12
fsspec==2021.5.0
h11==0.12.0
identify==2.2.4
idna==2.7
locket==0.2.1
memory-profiler==0.58.0
mo-dots==4.19.21059
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==3.93.20259
mo-logs==4.20.21059
moz-sql-parser==4.21.21059
multidict==5.1.0
nodeenv==1.6.0
numpy==1.20.2
pandas==1.2.4
partd==1.2.0
pre-commit==2.12.1
psutil==5.8.0
pycodestyle==2.7.0
pydantic~=1.8.2
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
starlette
toml==0.10.2
toolz==0.11.1
typing-extensions==3.7.4.3
urllib3==1.23
uvicorn==0.13.4
virtualenv==20.4.6
yarl==1.6.3
fastapi-users[mongodb]==8.0.0b2
motor~=2.5.1
python-dotenv~=0.19.0
pymongo~=3.12.1